// src/types/fx.ts
export interface FxPromptInputRef {
  /** Focus the input */
  focus: () => void;
  /** Clear the input */
  clear: () => void;
  /** Get current value */
  getValue: () => string;
  /** Set value */
  setValue: (value: string) => void;
  /** Access native DOM element */
  readonly nativeElement: HTMLElement | null;
  /** @deprecated Use nativeElement instead */
  getNativeElement: () => HTMLElement | null;
}